'#2371: Rename references to cryptocurrencies from CRIPTO to CRYPTO.
