+++ iped-app/resources/config/conf/metadataTypes.txt
@@ -1662,9 +1662,9 @@ Regex:BR_CPF = java.lang.String
 Regex:BR_PISPASEP = java.lang.String
 Regex:BR_TITULO_ELEITOR = java.lang.String
 Regex:CREDIT_CARD = java.lang.String
-Regex:CRIPTOCOIN_BITCOIN_ADDRESS = java.lang.String
-Regex:CRIPTOCOIN_MONERO = java.lang.String
-Regex:CRIPTO_POSSIBLE_SEED_PHRASE_EN = java.lang.String
+Regex:CRYPTOCOIN_BITCOIN_ADDRESS = java.lang.String
+Regex:CRYPTOCOIN_MONERO = java.lang.String
+Regex:CRYPTO_POSSIBLE_SEED_PHRASE_EN = java.lang.String
 Regex:EMAIL = java.lang.String
 Regex:IP = java.lang.String
 Regex:MAC_ADDRESS = java.lang.String
