+++ iped-app/resources/scripts/regex_validators/CryptoSeedPhraseValidator.js
