+++ iped-engine/src/main/java/iped/engine/task/regex/validator/crypto/LiteCoinAddressValidatorService.java
@@ -1,4 +1,4 @@
-package iped.engine.task.regex.validator.cripto;
+package iped.engine.task.regex.validator.crypto;
 
 import java.io.File;
 import java.util.Arrays;
@@ -6,21 +6,17 @@ import java.util.List;
 
 import iped.engine.task.regex.BasicAbstractRegexValidatorService;
 
-public class BitcoinAddressValidatorService extends BasicAbstractRegexValidatorService {
+public class LiteCoinAddressValidatorService extends BasicAbstractRegexValidatorService {
 
     private static final AltcoinBase58CheckValidator validator;
-
+    
     static {
         validator = new AltcoinBase58CheckValidator();
-        validator.setVersionForPrefix("1", (byte) 0);
-        validator.setVersionForPrefix("3", (byte) 5);
-        validator.setVersionForPrefix("5", (byte) 128);
-        validator.setVersionForPrefix("K", (byte) 128);
-        validator.setVersionForPrefix("L", (byte) 128);
-        validator.setVersionForPrefix("xpub", (byte) 4, (byte) 136, (byte) 178, (byte) 30);
-        validator.setVersionForPrefix("xpub", (byte) 4, (byte) 136, (byte) 173, (byte) 228);
+        validator.setVersionForPrefix("M", (byte) 50);
+        validator.setVersionForPrefix("L", (byte) 48);
+        validator.setVersionForPrefix("9", (byte) 22);
     }
-
+    
     private static int[] BECH32_CHARSET_REV = { -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1,
             -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1,
             -1, -1, -1, 15, -1, 10, 17, 21, 20, 26, 30, 7, 5, -1, -1, -1, -1, -1, -1, -1, 29, -1, 24, 13, 25, 9, 8, 23,
@@ -39,22 +35,14 @@ public class BitcoinAddressValidatorService extends BasicAbstractRegexValidatorS
 
     @Override
     public List<String> getRegexNames() {
-        return Arrays.asList("CRIPTOCOIN_BITCOIN_ADDRESS", "CRIPTOCOIN_BITCOIN_BIP38_ENC_PRIV_K",
-                "CRIPTOCOIN_BITCOIN_WIF_PRIV_K_UNC_PUB_K", "CRIPTOCOIN_BITCOIN_WIF_PRIV_K_COMP_PUB_K",
-                "CRIPTOCOIN_BITCOIN_BIP32_HD_XPRV_KEY", "CRIPTOCOIN_BITCOIN_BIP32_HD_XPUB_KEY");
+        return Arrays.asList("CRYPTOCOIN_LITECOIN");
     }
 
     @Override
     protected boolean validate(String hit) {
-        return validateBitcoinAddress(hit);
-    }
-
-    public boolean validateBitcoinAddress(String addr) {
-        if (addr.startsWith("bc1")) {
-            return bech32validator.bech32VerifyChecksum(addr);
+        if (hit.startsWith("ltc1")) {
+            return bech32validator.bech32VerifyChecksum(hit);
         }
-
-        return validator.validate(addr);
+        return validator.validate(hit);
     }
-
 }
